refactor commands directory.
